validators/tsv.js becomes module; fullTest.js uses tsv.validate()
